[zed-only] Fix functional tests
stable/zed was failing for the python-ironic-inspector-client
functional testing, which relied upon a sqllite database.

The inherent issue here is oslo.db 12.1.0 is available for
stable/zed, which disabled autocommit by default for compatability
with SQLAlchemy 2.0.

While one would think existing code in the branch would cover this,
the actual assertion of autocmmit needs to be done as early
as possible, so in this change we attempt to assert it before
any database access is started, which should allow stable/zed
functional testing to work. This is not required on later branches
as ironic-inspector's code was updated during the Antelope
development cycle to support SQLAlchemy 2.0.

Change-Id: I25bab2041e9d5b985a8978c5a32d4e46a0986e4a