Update the project status and move broken jobs to experimental
Since ironic-inspector is pending deprecation and does not see a lot of
changes, simply disable the jobs that do not pass. A release note is
issued about the status of the project.

Also: Re-adjust the memory footprint of the VMs since we use
tinycore linux on this job, *and* we were previously introsepcting
4 VMs at 1GB each. The count has been changed to two.

Change-Id: I11d10a9f83552b5875812fa65d87cc5c72081537