Don't fail if data already migrated
This is an attempt to soften the edges on the data migration command
where infrastructure operators may attempt to run the command multiple
times, even after, for example, swift no longer exists. This can cause
the command to error.

But the command doesn't need to error if the target *already* has the
needful data, as in the data was already migrated.

Change-Id: I9ce413b3b218b4655563153fde7131ff424bd7ee