Fix: ignore resolve error while looking up node
Log a warning instead of throwing an exception if an address of some node
cannot be resolved, so new nodes can be registered in ironic.

This solves the problem that if an address of any other node is not
resolvable, no new node can be registered, even if its own address is
resolvable.

Closes-bug: #2072544
Change-Id: If7f65bcac2bb23c148ba27d45166d4a978df1e79
Signed-off-by: Robert Hoffmann <robert.hoffmann@uhurutec.com>